simplified the product manipulation functionality

+++ ecommerce-be/src/main/java/com/ecommerce/ecommerce_be/controllers/CartController.java
@@ -4,10 +4,7 @@ import com.ecommerce.ecommerce_be.requests.AddNewCartItemRequest;
 import com.ecommerce.ecommerce_be.requests.CartContentsRequest;
 import com.ecommerce.ecommerce_be.services.CartService;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 public class CartController {
@@ -23,13 +20,13 @@ public class CartController {
     }
 
     @RequestMapping(value = "cartproduct", method = RequestMethod.POST)
-    public ResponseEntity<String> putProductInCart(@RequestBody AddNewCartItemRequest request) {
+    public ResponseEntity<String> productManipulation(@RequestParam(required = false, defaultValue = "false") Boolean remove,
+                                                   @RequestBody AddNewCartItemRequest request)
+    {
+        if(remove) {
+            return this.cartService.removeProductByRequest(request);
+        }
         return this.cartService.addCartProductByRequest(request);
-    }
-
-    @RequestMapping(value = "removeproduct", method = RequestMethod.POST)
-    public ResponseEntity<String> removeProductFromCart(@RequestBody AddNewCartItemRequest request) {
-        return this.cartService.removeProductByRequest(request);
     }
 
 }
